Add Licensing class  - parse entitlement cert for subscription metadata  - remove StubLicense  - remove unneeded license_key and references
